# frontend/src/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "API Zümm",
    "description" : "SIG apicole de gestion et de suivi de ruchers — contrat REST public. Les identifiants d'API (ex. getZummHoneyActualQuantity) ne se traduisent pas.",
    "license" : {
      "name" : "Propriétaire"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/calculateurs/valorisation" : {
      "get" : {
        "tags" : [ "calculateur-controller" ],
        "operationId" : "valorisation",
        "parameters" : [ {
          "name" : "kilos",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number"
          }
        }, {
          "name" : "prixKgEur",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "number"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Valorisation"
                }
              }
            }
          }
        }
      }
    },
    "/api/calculateurs/sirop" : {
      "get" : {
        "tags" : [ "calculateur-controller" ],
        "operationId" : "sirop",
        "parameters" : [ {
          "name" : "proportion",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "1:1"
          }
        }, {
          "name" : "litres",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Sirop"
                }
              }
            }
          }
        }
      }
    },
    "/api/calculateurs/refractometre" : {
      "get" : {
        "tags" : [ "calculateur-controller" ],
        "operationId" : "refractometre",
        "parameters" : [ {
          "name" : "indice",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "number"
          }
        }, {
          "name" : "temperatureC",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "number"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Refractometre"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Valorisation" : {
        "type" : "object",
        "properties" : {
          "kilos" : {
            "type" : "number"
          },
          "prixKgEur" : {
            "type" : "number"
          },
          "totalEur" : {
            "type" : "number"
          },
          "pots500g" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Sirop" : {
        "type" : "object",
        "properties" : {
          "proportion" : {
            "type" : "string"
          },
          "litres" : {
            "type" : "number"
          },
          "sucreKg" : {
            "type" : "number"
          },
          "eauL" : {
            "type" : "number"
          },
          "usage" : {
            "type" : "string"
          }
        }
      },
      "Refractometre" : {
        "type" : "object",
        "properties" : {
          "indice" : {
            "type" : "number"
          },
          "temperatureC" : {
            "type" : "number"
          },
          "indiceCorrige" : {
            "type" : "number"
          },
          "humiditePct" : {
            "type" : "number"
          },
          "verdict" : {
            "type" : "string"
          },
          "conformeNorme" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}